Debugging!
Debugging is a pain with this app. Since it uses gps I can only test it
using my phone and I have no console on a cell phone. I output to a text
view and I have to manually walk around to test + I need to always
update the app when I need to test a new fix.